salida y entrada inventario

+++ app/Http/Controllers/CategoriasController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Articulos;
 use App\Categorias;
 use Illuminate\Routing\Controller;
 use Illuminate\Http\Request;
@@ -17,7 +18,7 @@ class CategoriasController extends Controller
   {
     $listado_categorias = Categorias::all();
 
-    return view('categorias.index',compact('listado_categorias'));
+    return view('categorias.index', compact('listado_categorias'));
   }
 
   public function create()
@@ -27,11 +28,19 @@ class CategoriasController extends Controller
 
   public function store(Request $request)
   {
+    $validatedData = $request->validate([
+      'nombre' => 'required|unique:categorias|max:191',
+    ],[
+      'nombre.required' => 'El campo nombre es requerido',
+      'nombre.unique' => 'La categoria ingresada ya se encuentra registrada en la base de datos',
+      'nombre.max' => 'El campo nombre no debe superar 191 caracteres',
+    ]);
+
     $categoria = new Categorias();
     $categoria->nombre = $request->get('nombre');
     $categoria->save();
 
-    return redirect('categorias');
+    return redirect('categorias')->with('msj', 'Categoria agregada exitosamente');
   }
 
   public function edit($id)
@@ -43,20 +52,32 @@ class CategoriasController extends Controller
 
   public function update(Request $request)
   {
+    $validatedData = $request->validate([
+      'nombre' => 'required|max:191',
+    ],[
+      'nombre.required' => 'El campo nombre es requerido',
+      'nombre.max' => 'El campo nombre no debe superar 191 caracteres',
+    ]);
+
     $id = $request->get('id_categoria');
 
     $categoria = Categorias::findOrFail($id);
     $categoria->nombre = $request->get('nombre');
     $categoria->update();
 
-    return redirect('categorias');
+    return redirect('categorias')->with('msj', 'Categoria actualizada exitosamente');
   }
 
   public function destroy($id)
   {
-    Categorias::findOrFail($id)->delete();
-
-    return redirect('categorias');
+    //consultar si existen productos con esta categoria
+    $existProductos = Articulos::where('id_categoria', $id)->exists();
+    if ($existProductos) {
+      return redirect('categorias')->with('msjError', 'Existen productos relacionados a esta categoria');
+    } else {
+      Categorias::findOrFail($id)->delete();
+      return redirect('categorias')->with('msj', 'Categoria eliminada exitosamente');
+    }
   }
 
 }
